Deleting district now unbinds all candidates from that district

+++ src/main/java/voting/service/DistrictServiceImpl.java
@@ -40,7 +40,11 @@ public class DistrictServiceImpl implements DistrictService {
     @Transactional
     @Override
     public void deleteDistrict(Long id) {
-        districtRepository.delete(id);
+        District district = districtRepository.findOne(id);
+        if (district != null) {
+            district.removeAllCandidates();
+            districtRepository.delete(id);
+        }
     }
 
     @Override
